# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Veterinary Management
  description: Veterinary Management
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
    description: "Local server"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication using OAuth2 Bearer token.

        **How to authenticate:**
        1. POST to `/loginJWT` with `administrator_id` and `password`
        2. Receive JWT token in response
        3. Include token in Authorization header: `Bearer <token>`

        **Available scopes:**
        - `RECEPTIONIST` - Clinic receptionist staff (manage visits, view schedules)
        - `CLINIC_MANAGER` - Clinic manager (full system access)
        - `VETERINARIAN` - Veterinarian (manage own schedule, view assigned visits)
        - `PET_OWNER` - Pet owner (view own pets and visits)
  parameters:
...
    medication_id:
      name: medication_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/MedicationId'
...
    invoice_id:
      name: invoice_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/InvoiceId'
...
  requestBodies:
...
    MedicationBatchInput:
      description: "Medication batch input"
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              lot_number:
                type: integer
                format: int64
              received_date:
                $ref: '#/components/schemas/LocalDate'
              expiry_date:
                $ref: '#/components/schemas/LocalDate'
              initial_quantity:
                type: integer
              current_quantity:
                type: integer
              purchase_price_per_unit:
                type: number
                format: double
...
    InvoiceInput:
      description: "Invoice input"
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              visit_id:
                $ref: '#/components/schemas/VisitId'
    PaymentInput:
      description: "Payment input"
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              payment_method:
                $ref: '#/components/schemas/PaymentMethod'
              transaction_reference:
                $ref: '#/components/schemas/LocalDate'
  schemas:
...
    LocalDate:
      type: string
      format: date
      description: "Local date in the format YYYY-MM-DD"
      default: "2025-01-01"
...
    MedicationId:
      type: integer
      format: int64
      description: "Unique identifier for an Medication"
    BatchId:
      type: integer
      format: int64
      description: "Unique identifier for an Batch"
...
    VisitId:
      type: integer
      format: int64
      description: "Unique identifier for a Visit"
...
    PetId:
      type: integer
      format: int64
      description: "Unique identifier for a Pet"
    PetOwnerId:
      type: integer
      format: int64
      description: "Unique identifier for a Pet Owner"
...
    Medication:
      type: object
      properties:
        medication_id:
          $ref: '#/components/schemas/MedicationId'
        name:
          type: string
        active_ingredient:
          type: string
        dosage_unit:
          type: integer
        unit_price:
          type: number
          format: double
...
    MedicationBatch:
      type: object
      properties:
        batch_id:
          $ref: '#/components/schemas/BatchId'
        medication_id:
          $ref: '#/components/schemas/MedicationId'
        lot_number:
          type: integer
          format: int64
        received_date:
          $ref: '#/components/schemas/LocalDate'
        expiry_date:
          $ref: '#/components/schemas/LocalDate'
        initial_quantity:
          type: integer
        current_quantity:
          type: integer
        purchase_price_per_unit:
          type: number
          format: double
...
    InvoiceId:
      type: integer
      format: int64
      description: "Unique identifier for an Invoice"
    Invoice:
      type: object
      properties:
        invoice_id:
          $ref: '#/components/schemas/InvoiceId'
        visit_id:
          $ref: '#/components/schemas/VisitId'
        invoice_date:
          $ref: '#/components/schemas/LocalDate'
        amount:
          type: number
          format: double
        due_date:
          $ref: '#/components/schemas/LocalDate'
        paid:
          type: boolean
    PaymentMethod:
      type: string
      enum: [ CASH, CREDIT_CARD, DEBIT_CARD, BANK_TRANSFER ]
      description: "Method of payment"
...
paths:
...
  /medications/{medication_id}:
    get:
      tags: [ Medications ]
      summary: "Get a medication"
      security:
        - bearerAuth: [ CLINIC_MANAGER ]
      parameters:
        - $ref: '#/components/parameters/medication_id'
      responses:
        '200':
          description: "Medication found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medication'
        '404':
          description: "Medication not found"
  /medications/{medication_id}/medicationBatches:
    post:
      tags: [ Medications ]
      summary: "Create a new batch for a medication"
      security:
        - bearerAuth: [CLINIC_MANAGER]
      parameters:
        - $ref: '#/components/parameters/medication_id'
      requestBody:
        $ref: '#/components/requestBodies/MedicationBatchInput'
      responses:
        '201':
          description: "Medication batch created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationBatch'
        '400':
          description: "Invalid input"
        '404':
          description: "Medication not found"
    get:
      tags: [ Medications ]
      summary: "Get all medication batches from a Medication"
      security:
        - bearerAuth: [ CLINIC_MANAGER ]
      parameters:
        - $ref: '#/components/parameters/medication_id'
      responses:
        '200':
          description: "Medication batches found"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MedicationBatch'
        '404':
          description: "Medication not found"
...
  /invoices:
    post:
      tags: [ Billing & Payments ]
      summary: "Generate an invoice for a visit"
      security:
        - bearerAuth: [ RECEPTIONIST, CLINIC_MANAGER ]
      requestBody:
        $ref: '#/components/requestBodies/InvoiceInput'
      responses:
        '201':
          description: "Invoice generated"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '404':
          description: "Visit not found"
        '409':
          description: "Invoice already exists for this visit"
    get:
      tags: [ Billing & Payments ]
      summary: "Get all invoices"
      security:
        - bearerAuth: [ RECEPTIONIST, CLINIC_MANAGER ]
      parameters:
        - name: startDate
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/LocalDate'
        - name: endDate
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/LocalDate'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ PAID, UNPAID ]
        - name: petOwnerId
          in: query
          required: false
          schema:
             $ref: '#/components/schemas/PetOwnerId'
        - name: petId
          in: query
          required: false
          schema:
             $ref: '#/components/schemas/PetId'
      responses:
        '200':
          description: "Invoices found"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
        '404':
          description: "Invoice not found"
  /invoices/{invoice_id}:
    get:
      tags: [ Billing & Payments ]
      summary: "Get an invoice"
      security:
        - bearerAuth: [ RECEPTIONIST, CLINIC_MANAGER ]
      parameters:
        - $ref: '#/components/parameters/invoice_id'
      responses:
        '200':
          description: "Invoice found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          description: "Invalid input"
        '404':
          description: "Invoice not found"
    delete:
      tags: [ Billing & Payments ]
      summary: "Delete an invoice"
      security:
        - bearerAuth: [ RECEPTIONIST, CLINIC_MANAGER ]
      parameters:
        - $ref: '#/components/parameters/invoice_id'
      responses:
        '204':
          description: "Invoice deleted"
        '404':
          description: "Invoice not found"
  /invoices/{invoice_id}/pay:
    post:
      tags: [ Billing & Payments ]
      summary: "Pay an invoice"
      security:
        - bearerAuth: [ RECEPTIONIST, CLINIC_MANAGER ]
      parameters:
        - $ref: '#/components/parameters/invoice_id'
      requestBody:
        $ref: '#/components/requestBodies/PaymentInput'
      responses:
        '200':
          description: "Invoice paid"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          description: "Invalid input"
        '404':
          description: "Invoice not found"
        '409':
          description: "Invoice already paid"
...